apps/api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic[email]==2.5.0
pymongo==4.6.0
motor==3.3.2
python-multipart==0.0.6
PyMuPDF==1.23.8
python-docx==1.1.0
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
pyyaml==6.0.1
selectolax==0.3.17
feedparser==6.0.10

